+++ changelog.md
@@ -3,6 +3,7 @@
 * Optimized and threaded shape generation to avoid spikes at super complex models
 * Beds now allow respawning
 * New worlds can now be copied without data loss
+* Destroying furniture now kicks off passengers
 
 # 0.0.7
 
@@ -62,7 +63,5 @@ Initial release
 * Redstone input shall trigger right click
 * Multi craft intransparent and hard to use (E.g. "hold shift, hold space")
 * On flat elements (sprites mostly) the direction is picked oddly
-* Check if the block still exists when interacting
 * Better dismounting
 * Filter likes button and config
-* Why no leaves?